Update FBShipIt description, add adeira/shipit (#60)
A long time ago, we ported FBShipIt to JavaScript, so it's easier for us to run it. It helps us to export commits from our monorepo ever since. Even `adeira/shipit` is exported via itself. Recently, we open-sourced it, and I believe it could be useful even for other devs interested in doing the same.